(chore): move hash functions to SDK
